# server/api/gen/go/admin/auth_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             (unknown)
// source: admin/auth.proto

package admin

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
	emptypb "google.golang.org/protobuf/types/known/emptypb"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationAuthServiceGetUserInfo = "/admin.AuthService/GetUserInfo"
const OperationAuthServiceGetUserMenu = "/admin.AuthService/GetUserMenu"
const OperationAuthServiceGetUserProfile = "/admin.AuthService/GetUserProfile"
const OperationAuthServiceLogin = "/admin.AuthService/Login"
const OperationAuthServiceSendUpdatePhoneCode = "/admin.AuthService/SendUpdatePhoneCode"
const OperationAuthServiceUpdateUserPhone = "/admin.AuthService/UpdateUserPhone"
const OperationAuthServiceUpdateUserProfile = "/admin.AuthService/UpdateUserProfile"
const OperationAuthServiceUpdateUserPwd = "/admin.AuthService/UpdateUserPwd"

type AuthServiceHTTPServer interface {
	// GetUserInfo 获取已经登录的用户的数据
	GetUserInfo(context.Context, *emptypb.Empty) (*UserInfo, error)
	// GetUserMenu 获取已经登录的用户菜单
	GetUserMenu(context.Context, *emptypb.Empty) (*TreeRouteResponse, error)
	// GetUserProfile 获取个人中心用户信息
	GetUserProfile(context.Context, *emptypb.Empty) (*UserProfileForm, error)
	// Login 登录
	Login(context.Context, *LoginRequest) (*LoginResponse, error)
	// SendUpdatePhoneCode 发送手机号验证码
	SendUpdatePhoneCode(context.Context, *SendUpdatePhoneCodeForm) (*emptypb.Empty, error)
	// UpdateUserPhone 修改个人中心手机号
	UpdateUserPhone(context.Context, *UpdatePhoneForm) (*emptypb.Empty, error)
	// UpdateUserProfile 修改个人中心用户信息
	UpdateUserProfile(context.Context, *UserProfileForm) (*emptypb.Empty, error)
	// UpdateUserPwd 修改个人中心密码
	UpdateUserPwd(context.Context, *UpdatePwdForm) (*emptypb.Empty, error)
}

func RegisterAuthServiceHTTPServer(s *http.Server, srv AuthServiceHTTPServer) {
	r := s.Route("/")
	r.POST("/api/admin/auth/login", _AuthService_Login0_HTTP_Handler(srv))
	r.GET("/api/admin/auth/userInfo", _AuthService_GetUserInfo0_HTTP_Handler(srv))
	r.GET("/api/admin/auth/menu", _AuthService_GetUserMenu0_HTTP_Handler(srv))
	r.GET("/api/admin/auth/userProfile", _AuthService_GetUserProfile0_HTTP_Handler(srv))
	r.PUT("/api/admin/auth/update/userProfile", _AuthService_UpdateUserProfile0_HTTP_Handler(srv))
	r.POST("/api/admin/auth/send/update/phone", _AuthService_SendUpdatePhoneCode0_HTTP_Handler(srv))
	r.PUT("/api/admin/auth/update/phone", _AuthService_UpdateUserPhone0_HTTP_Handler(srv))
	r.PUT("/api/admin/auth/update/pwd", _AuthService_UpdateUserPwd0_HTTP_Handler(srv))
}

func _AuthService_Login0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LoginRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceLogin)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.Login(ctx, req.(*LoginRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LoginResponse)
		return ctx.Result(200, reply)
	}
}

func _AuthService_GetUserInfo0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceGetUserInfo)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserInfo(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UserInfo)
		return ctx.Result(200, reply)
	}
}

func _AuthService_GetUserMenu0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceGetUserMenu)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserMenu(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*TreeRouteResponse)
		return ctx.Result(200, reply)
	}
}

func _AuthService_GetUserProfile0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in emptypb.Empty
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceGetUserProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUserProfile(ctx, req.(*emptypb.Empty))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*UserProfileForm)
		return ctx.Result(200, reply)
	}
}

func _AuthService_UpdateUserProfile0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UserProfileForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceUpdateUserProfile)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateUserProfile(ctx, req.(*UserProfileForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _AuthService_SendUpdatePhoneCode0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendUpdatePhoneCodeForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceSendUpdatePhoneCode)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendUpdatePhoneCode(ctx, req.(*SendUpdatePhoneCodeForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _AuthService_UpdateUserPhone0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdatePhoneForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceUpdateUserPhone)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateUserPhone(ctx, req.(*UpdatePhoneForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

func _AuthService_UpdateUserPwd0_HTTP_Handler(srv AuthServiceHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in UpdatePwdForm
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationAuthServiceUpdateUserPwd)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.UpdateUserPwd(ctx, req.(*UpdatePwdForm))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*emptypb.Empty)
		return ctx.Result(200, reply)
	}
}

type AuthServiceHTTPClient interface {
	// GetUserInfo 获取已经登录的用户的数据
	GetUserInfo(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *UserInfo, err error)
	// GetUserMenu 获取已经登录的用户菜单
	GetUserMenu(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *TreeRouteResponse, err error)
	// GetUserProfile 获取个人中心用户信息
	GetUserProfile(ctx context.Context, req *emptypb.Empty, opts ...http.CallOption) (rsp *UserProfileForm, err error)
	// Login 登录
	Login(ctx context.Context, req *LoginRequest, opts ...http.CallOption) (rsp *LoginResponse, err error)
	// SendUpdatePhoneCode 发送手机号验证码
	SendUpdatePhoneCode(ctx context.Context, req *SendUpdatePhoneCodeForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// UpdateUserPhone 修改个人中心手机号
	UpdateUserPhone(ctx context.Context, req *UpdatePhoneForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// UpdateUserProfile 修改个人中心用户信息
	UpdateUserProfile(ctx context.Context, req *UserProfileForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
	// UpdateUserPwd 修改个人中心密码
	UpdateUserPwd(ctx context.Context, req *UpdatePwdForm, opts ...http.CallOption) (rsp *emptypb.Empty, err error)
}

type AuthServiceHTTPClientImpl struct {
	cc *http.Client
}

func NewAuthServiceHTTPClient(client *http.Client) AuthServiceHTTPClient {
	return &AuthServiceHTTPClientImpl{client}
}

// GetUserInfo 获取已经登录的用户的数据
func (c *AuthServiceHTTPClientImpl) GetUserInfo(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*UserInfo, error) {
	var out UserInfo
	pattern := "/api/admin/auth/userInfo"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAuthServiceGetUserInfo))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserMenu 获取已经登录的用户菜单
func (c *AuthServiceHTTPClientImpl) GetUserMenu(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*TreeRouteResponse, error) {
	var out TreeRouteResponse
	pattern := "/api/admin/auth/menu"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAuthServiceGetUserMenu))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUserProfile 获取个人中心用户信息
func (c *AuthServiceHTTPClientImpl) GetUserProfile(ctx context.Context, in *emptypb.Empty, opts ...http.CallOption) (*UserProfileForm, error) {
	var out UserProfileForm
	pattern := "/api/admin/auth/userProfile"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationAuthServiceGetUserProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// Login 登录
func (c *AuthServiceHTTPClientImpl) Login(ctx context.Context, in *LoginRequest, opts ...http.CallOption) (*LoginResponse, error) {
	var out LoginResponse
	pattern := "/api/admin/auth/login"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthServiceLogin))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendUpdatePhoneCode 发送手机号验证码
func (c *AuthServiceHTTPClientImpl) SendUpdatePhoneCode(ctx context.Context, in *SendUpdatePhoneCodeForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/auth/send/update/phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthServiceSendUpdatePhoneCode))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateUserPhone 修改个人中心手机号
func (c *AuthServiceHTTPClientImpl) UpdateUserPhone(ctx context.Context, in *UpdatePhoneForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/auth/update/phone"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthServiceUpdateUserPhone))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateUserProfile 修改个人中心用户信息
func (c *AuthServiceHTTPClientImpl) UpdateUserProfile(ctx context.Context, in *UserProfileForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/auth/update/userProfile"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthServiceUpdateUserProfile))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// UpdateUserPwd 修改个人中心密码
func (c *AuthServiceHTTPClientImpl) UpdateUserPwd(ctx context.Context, in *UpdatePwdForm, opts ...http.CallOption) (*emptypb.Empty, error) {
	var out emptypb.Empty
	pattern := "/api/admin/auth/update/pwd"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationAuthServiceUpdateUserPwd))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "PUT", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}